Replace Grand Theft Auto
The biggest issue with the existing entry for GTA is the CHD. I haven't been able to reproduce with a matching MD5. The closest I found is Redump's dump of the Director's Cut version of the game. The file tree is nearly identical. This PR replaces existing info in the DAT with info generated from extracting data from this disc: http://redump.org/disc/30124/ + a CHD made from a bin/cue of the disc.