Add The Colonel's Bequest
## The Colonel's Bequest (1989) (Sierra On-Line)

Installed from 4 3.5" floppy disks matching [The Good Old Days ID: 171](https://www.goodolddays.net/diskimages/id%2C171/).

The DOSC auto starts and contains `RESOURCE.CFG` which configures graphics to EGA/VGA, sound to AdLib, and enables keyboard/mouse.

It also contains a copy protection patch in IPS format that I made based on the instructions near the bottom of [this page](https://www.benshoof.org/blog/case-cracked). Long story short, the most widespread crack for much of the game's existence had the unfortunate side effect of completely nuking the game's RNG. So all of the random events in the game either never happened or always happened. This patch should remove the copy protection but leave the RNG intact.

Install disks are added as an installation variant.

The GOG version of the game is added as a digital variant. The DOSC matches the main entry, including the copy protection patch.