Add Computer Underground
## Game (Year) (Developer)

### Source Media

- Installed from Redump-verified CD-ROM: [XXXX](http://redump.org/disc/XXXX/)
- Installed from X X.X" floppy disks matching The Good Old Days ID: [XXXX](https://www.goodolddays.net/en/diskimages/?id=XXXX)
- Matches TDC Entry: `XXXX`

### Installation Options

- Options

How to run the DOSZ without the DOSC, if necessary.

Configuration is contained in the DOSC instead of the DOSZ, so installation options do not matter. To play the game without the DOSC, users can run `XXXX` to produce `XXXX`.

### Official Patches

Installed the official XXXX patch found [here](XXXX). Patch information.

No official patch is necessary.

### DOSC Contents

- Auto start
- Configuration file
- Copy protection patch
- MT-32 option

### Installation Variants

Install disks matching TGOD ID XXXX have been added as an installation variant.

The Redump-verified CD-ROM has been added as an installation variant.

The CD is required to play, so no installation variant is provided.

### Digital Store Variants

The GOG version is provided as a variant. DOSC information.

The Steam version is provided as a variant. DOSC information.

The GOG and Steam versions match and are provided as a variant. DOSC information.

The GOG/Steam version matches the base, so no variant is provided.

The game is not available on GOG or Steam.

### Unofficial Patch Variants

XXXX patch has been added as a miscellaneous variant. Patch information.

No unofficial patches were found.