Add Cyberia Series
## Cyberia (1994) (Xatrix Entertainment)

### Source Media

- Installed from Redump-verified CD-ROM: [Cyberia (USA)](http://redump.org/disc/22394/)

### Installation Options

- Default Directory

### Official Patches

No official patch is necessary.

### DOSC Contents

- Mounts disc
- Auto start

### Installation Variants

The CD is required to play, so no installation variant is provided.

### Digital Store Variants

I tried adding the GOG variant, but I'd need to be able to mount a directory as a CD. Trying to start the game gives an error about not being able to detect the CD.

### Unofficial Patch Variants

No unofficial patches were found.

***

## Cyberia 2 - Resurrection (1995) (Xatrix Entertainment)

### Source Media

- Files extracted from Redump-verified CD-ROMs: [Cyberia 2 - Resurrection (USA) (Disc 1) (Rerelease)](http://redump.org/disc/45858/) and [Cyberia 2 - Resurrection (USA) (Disc 2) (Rerelease)](http://redump.org/disc/45859/)

### Installation Options

- Default directory

### Official Patches

No official patch is necessary.

### DOSC Contents

- Auto start

### Installation Variants

The Redump-verified CD-ROMs have been added as an installation variant.

### Digital Store Variants

The [GOG version](https://www.gog.com/en/game/cyberia_2_resurrection) is provided as a variant.

### Unofficial Patch Variants

No unofficial patches were found.