Add Putt-Putt Series
## Putt-Putt Joins the Parade (1992) (Humongous Entertainment)

### Source Media

- Contents extracted from Redump-verified CD-ROM: [Putt-Putt Joins the Parade (USA)](http://redump.org/disc/43610/)

### Installation Options

- SoundBlaster Pro selected

### Official Patches

No official patch is necessary.

### DOSC Contents

- Mounts disc
- Auto start
- Configuration file

### Installation Variants

The CD is required to play, so no installation variant is provided.

### Digital Store Variants

The DOS version of the game is not available on GOG or Steam.

### Unofficial Patch Variants

No unofficial patches were found.

***

## Putt-Putt's Fun Pack (1993) (Humongous Entertainment)

### Source Media

- Contents extracted from Redump-verified CD-ROM: [Putt-Putt's Fun Pack (USA) (Rerelease)](http://redump.org/disc/51069/)

### Installation Options

- SoundBlaster Pro selected

### Official Patches

No official patch is necessary.

### DOSC Contents

- Mounts disc
- Auto start
- Configuration file

### Installation Variants

The CD is required to play, so no installation variant is provided.

### Digital Store Variants

The DOS version of the game is not available on GOG or Steam.

### Unofficial Patch Variants

No unofficial patches were found.

***

## Putt-Putt Goes to the Moon (1993) (Humongous Entertainment)

### Source Media

- Contents extracted from Redump-verified CD-ROM: [Putt-Putt Goes to the Moon (USA)](http://redump.org/disc/50326/)

### Installation Options

- SoundBlaster Pro selected

### Official Patches

No official patch is necessary.

### DOSC Contents

- Mounts disc
- Auto start
- Configuration file

### Installation Variants

The CD is required to play, so no installation variant is provided.

### Digital Store Variants

The DOS version of the game is not available on GOG or Steam.

### Unofficial Patch Variants

No unofficial patches were found.

***

## Putt-Putt & Fatty Bear's Activity Pack (1994) (Humongous Entertainment)

### Source Media

- Contents extracted from Redump-verified CD-ROM: [Putt-Putt & Fatty Bear's Activity Pack (USA)](http://redump.org/disc/51050/)

### Installation Options

- SoundBlaster Pro selected

### Official Patches

No official patch is necessary.

### DOSC Contents

- Mounts disc
- Auto start
- Configuration file

### Installation Variants

The CD is required to play, so no installation variant is provided.

### Digital Store Variants

The DOS version of the game is not available on GOG or Steam.

### Unofficial Patch Variants

No unofficial patches were found.